Тесты вынесены в отдельные файлы по endpoint'ам. Добавлены новые тесты.
Обновлены зависимости.
