+++ requirements.txt
@@ -10,6 +10,8 @@ databases==0.4.1
 dnspython==2.1.0
 email-validator==1.1.2
 fastapi==0.63.0
+freezegun==1.1.0
+greenlet==1.0.0
 h11==0.12.0
 idna==3.1
 iniconfig==1.1.1
@@ -24,6 +26,8 @@ pydantic==1.7.3
 PyJWT==2.0.1
 pyparsing==2.4.7
 pytest==6.2.2
+python-dateutil==2.8.1
+six==1.15.0
 SQLAlchemy==1.3.23
 starlette==0.13.6
 toml==0.10.2
