+++ src/auth/tests/utils.py
@@ -1,0 +1,9 @@
+import datetime
+
+
+def create_fake_datetime(date):
+    class FakeDateTime(datetime.datetime):
+        @classmethod
+        def utcnow(cls):
+            return date
+    return FakeDateTime
